requirements.txt
vanna>=0.8.0
pymysql>=1.1.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pandas>=2.0.0
plotly>=5.18.0
ollama>=0.1.0
google-genai>=1.0.0
chromadb>=0.4.0
